refactor(frontend): split contexts/hooks out of component files
Fast refresh breaks when a module exports both components and
non-component values (contexts, hooks, constants). Move contexts and
shared hooks to dedicated modules:

- context/LiffContext: LiffContext moved out of LiffProvider.jsx
- theme/useColorMode: ColorModeContext + hook moved out of ColorModeContext.jsx
- pages/Rankings/hooks: data hooks moved out of the three chart components

Also cleans up unused vars flagged by the same pass:
- Race/index: drop status prop now that it's unused post raceFinished removal
- Admin/Messages: drop useless initializers on switch-reassigned locals
- AutoSettings: bare catch for unused err
