+++ frontend/src/context/LiffContext.js
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+export const LiffContext = createContext({
+  ready: false,
+  loggedIn: false,
+  isAdmin: false,
+  profile: null,
+  liffContext: {},
+  login: () => {},
+  logout: () => {},
+});
