+++ frontend/src/context/useLiff.js
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { LiffContext } from "./LiffProvider";
+import { LiffContext } from "./LiffContext";
 
 export default function useLiff() {
   return useContext(LiffContext);
