+++ frontend/src/pages/Rankings/ChatLevelChart.jsx
@@ -1,22 +1,6 @@
-import { useMemo } from "react";
-import useAxios from "axios-hooks";
 import RankingBarChart from "./RankingBarChart";
 import { RANK_COLORS } from "./OverviewCard";
-
-export function useChatLevelData() {
-  const [{ data, loading }] = useAxios("/api/chat-levels/rankings");
-
-  const rows = useMemo(() => {
-    if (!data) return [];
-    return data.map(d => ({
-      displayName: d.displayName,
-      value: d.experience,
-      level: d.level,
-    }));
-  }, [data]);
-
-  return { rows, loading, topEntry: rows[0], count: rows.length };
-}
+import { useChatLevelData } from "./hooks";
 
 export default function ChatLevelChart() {
   const { rows } = useChatLevelData();
