+++ frontend/src/pages/Rankings/GodStoneChart.jsx
@@ -1,21 +1,6 @@
-import { useMemo } from "react";
-import useAxios from "axios-hooks";
 import RankingBarChart from "./RankingBarChart";
 import { RANK_COLORS } from "./OverviewCard";
-
-export function useGodStoneData() {
-  const [{ data, loading }] = useAxios("/api/god-stone/rankings");
-
-  const rows = useMemo(() => {
-    if (!data) return [];
-    return data.map(d => ({
-      displayName: d.displayName,
-      value: d.amount,
-    }));
-  }, [data]);
-
-  return { rows, loading, topEntry: rows[0], count: rows.length };
-}
+import { useGodStoneData } from "./hooks";
 
 export default function GodStoneChart() {
   const { rows } = useGodStoneData();
