+++ frontend/src/theme/useColorMode.js
@@ -1,0 +1,10 @@
+import { createContext, useContext } from "react";
+
+export const ColorModeContext = createContext({
+  mode: "light",
+  toggleColorMode: () => {},
+});
+
+export function useColorMode() {
+  return useContext(ColorModeContext);
+}
